Rename moduleData file to use kebab-case and update it's references.
